requirements.txt
# War Image Classification - Dependencies
# =========================================

# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine learning utilities
scikit-learn>=1.3.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.7.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Configuration
PyYAML>=6.0

# Kaggle data download
kagglehub>=0.2.0

# Environment variables
python-dotenv>=1.0.0

